we do a little refactoring + correct testing methods
