+++ solib/__init__.py
@@ -1,0 +1,4 @@
+import logging
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
