+++ solib/protocols/variants/consultancy.py
@@ -1,9 +1,10 @@
 from solib.datatypes import Answer, Question
 from solib.protocols.consultancy import Consultant, Client, Consultancy
-from solib.protocols.common import Judge, Transcript
+from solib.protocols.abstract import Judge, Transcript
+
 
 class OpenConsultancy(Consultancy):
-    
+
     async def choose_answer(self, question: Question, **kwargs) -> Answer:
         auxillary_judge = Judge(model=self.consultant.model)
         transcript = Transcript(question)
